### 2019-12-12 #1950 @kyleniemeyer
Hi @kbarnhart, sorry for two requests today, but could you edit this one as well? With the focus on remote observations I thought it might be in your wheelhouse.
