+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). 
+Unfortunatly, this project does not adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) because the author discovered the semantic versionning after his first releases.. 
 
 ## [Unreleased]
 ### added
