+++ museotoolbox/__init__.py
@@ -14,4 +14,4 @@
 # =============================================================================
 from __future__ import absolute_import
 from . import learn_tools, vector_tools, raster_tools, cross_validation, datasets, stats  # , charts
-__version__ = "1.4.1"
+__version__ = "0.10.01"
