### 2019-12-13 536cd72
Enhance getSamplesFromROI to get only raster values from vector, (no need to put a field from vector file)
