+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). 
-Unfortunatly, this project does not adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) because the author discovered the semantic versionning after his first releases.. 
+This project strictly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.12] - 2019-12-13
 ### changed
