### 2019-12-13 892a7ae
getSamplesFromROI return list of available fields if wrong field given. rasterMath convert np.nan value to nodata value (if numpy >= 1.17)
