### 2019-12-13 dfdf213
- Minor fix when using learnAndPredict with an outside customized function - Better management fo cross-validation in learnAndPredict - Fix minor bug using False or None value with cv in learnAndPredict - Add an option to use SFS without writing each best model on the disk.
