### 2019-12-13 5304cc2
added - Move some functions from vector_tools to raster_tools, functions are anyway still available from vector_tools
changed
- learnAndPredict manages int value for cross-validation by using RandomStratifiedKFold
- Enhance blocksize management for rasterMath
- Move command line code in _cli folder
