### 2019-12-13 fdc7863
v0.11.02 changed - Improvements of rasterMath - customBlockSize defines now the same block size for window reading and for the output - add seed parameter (to set a random generator) in getRandomBlock() - add getRasterParameters() and customRasterParameters() function.
