+++ museotoolbox/__init__.py
@@ -13,4 +13,4 @@
 # @git:     www.github.com/nkarasiak/MuseoToolBox
 # =============================================================================
 from . import learn_tools, vector_tools, raster_tools, cross_validation, datasets, stats  # , charts
-__version__ = "1.6.3"
+__version__ = "0.11.02"
