### 2019-12-13 #1950 @nkarasiak
Hi, just to inform you that to respect semantic versioning I changed how I named my version. So the current version to review is now v0.12.
Kind regards,
Nicolas.
