### 2019-12-13 #1950 @kbarnhart
@kyleniemeyer I can handle this one.
