### 2019-12-16 5b252c2
update changelog and version number
