+++ changelog.txt
@@ -1,67 +1,0 @@
-=============
-CHANGELOG
-=============
-2.0.1
------
-- add zonal_stats example
-- update docstring in ai
-- modify SequentialFeatureSelection to have same fit method as SuperLearner
-
-2.0
------
-Update RasterMath and readme file.
-
-2.0b1
------
-Lot of changes to satisfy python conventions.
-
-Some folders have changed name :
-    - raster_tools and vector_tools > processing
-    - learn_tools > ai
-
-some functions have changed name : 
-    - getSamplesFromROI > extract_values
-    - historicalMap > load_historical_data
-    - getDistanceMatrix > get_distance_matrix
-
-classes now always begin with a capital case : 
-    - learnAndPredict > SuperLearner
-    - rasterMath > RasterMath
-    - sequentialFeatureSelection > SequentialFeatureSelection 
-
-- Fix bug #7 : getSamplesFromROI (nowd extract_ROI) now extracts ROI values using by default memory. If it fails, it will create a temporary raster on disk then delete it when finished. 
-- Remove cli
-
-1.6.6
-------
-* getSamplesFromROI return list of available fields if wrong field given.
-* rasterMath convert np.nan value to nodata value (if numpy >= 1.17)
-
-1.6.5
-------
-* Minor fix when using learnAndPredict with an outside customized function
-* Better management fo cross-validation in learnAndPredict
-** Fix minor bug using False or None value with cv in learnAndPredict
-* Add an option to use SFS without writing each best model on the disk.
-
-1.6.4
------
-* learnAndPredict manages int value for cross-validation by using RandomStratifiedKFold
-* Enhance blocksize management for rasterMath
-* Move command line code in _cli folder
-* Move some functions from vector_tools to raster_tools, functions are anyway still available from vector_tools
-
-1.6.3
-------
-* Improvements of rasterMath
-** customBlockSize defines now the same block size for window reading and for the output
-** add seed parameter (to set a random generator) in getRandomBlock()
-** add getRasterParameters() and customRasterParameters() function.
-
-
-1.6.2
-------
-* update rasterMath to generate by default a 256*256 raster block size.
-* update rasterMath to prevent bug if user has osgeo/gdal<2.1.
-* prevent bug when in rasterMath if processor has only 1 core.
-* minor fixes and doc update
