### 2019-12-16 #1950 @nkarasiak
Hello,
Here is a list of the three potential reviewers I identified :

- MES1605 (satellite images and and classification)
- cmillion (image processing & python)
- OopsAllErrors (remote sensing & image processing)

Kind regards,
Nicolas Karasiak.
