### 2019-12-17 10b976c
fix bug in RasterMath / get_block
