### 2019-12-17 7d8121c
fix bug in RasterMath / get_block
