### 2019-12-17 #1950 @kbarnhart
Thanks for the recommendation @cmillion

:wave: @mollenburger would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue that I use to coordinate reviewers. Once there are sufficient reviewers (usually 2-3), I will open a new issue where the review will take place.

If you are not able, please let me know if you have recommendations for other reviewers.
