### 2019-12-20 #1978 @whedon
[REVIEW]: Museo ToolBox : a python library for remote sensing including a new way to handle rasters.
**Submitting author:** @nkarasiak (<a href="http://orcid.org/0000-0002-1558-0816">Nicolas Karasiak</a>)
**Repository:** <a href="https://github.com/nkarasiak/MuseoToolBox" target ="_blank">https://github.com/nkarasiak/MuseoToolBox</a>
**Version:** v0.13.0
