### 2019-12-20 #1950 @kbarnhart
I'm going to transition this submission from pre-review to review with two reviewers.

Thanks all for your enthusiasm for finding and recommending reviewers. I've started the formal review over at #1978.

For those of you who are not yet listed on our [list of potential reviewers](https://bit.ly/joss-reviewers) for future submissions, please consider filling out the 1-minute form at [this link](https://joss.theoj.org/reviewer-signup.html). This helps editors and authors find appropriate reviewers.
