### 2019-12-20 #1978 @kbarnhart
:wave: @nkarasiak, @cmillion, @mollenburger this is the review thread for the paper. All of our communications will happen here from now on.

All reviewers have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time. I understand that it is late December and many people may be taking vacation and/or spending time away from work.  We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
