+++ CHANGELOG.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased : 0.12.1rc1] - 
+## [Unreleased : 0.12.1-beta.1] - 
 
 ### Changed
 
 - SequentialFeatureSelection parameters order Changed. *scoring* is now before *standardize*.
 - Fix bug in get_block() and get_random_block() which returned the same block each time due to new method.
+- Update doc for load_historical_data()
+- Fix bug with nodata in RasterMath when output is of float type
 
 ## [0.12] - 2019-12-13
 
