+++ CHANGELOG.md
@@ -84,3 +84,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update rasterMath to prevent bug if user has osgeo/gdal version is lower than 2.1.
 - prevent bug when in rasterMath if processor has only 1 core.
 - minor fixes and doc update
+
+### Added
+- new branch spatial added 
