+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased : 0.12.1-beta.1] - 
+## [Unreleased : 0.12.1-beta.2]
+
+### Added
+
+- n_jobs for RasterMath
+- spatial filtering for RasterMath
+
+## [0.12.1-beta.1] - 2020-01-16
 
 ### Changed
 
