### 2020-01-20 #1978 @kbarnhart
:wave: @cmillion, @mollenburger A friendly reminder to complete your reviews. If you have an estimated timeline, please let me know.

@mollenburger could you please confirm receipt.

As always, if you have any questions, please feel free to ping me here.

Thank you for being willing to review for JOSS.
