### 2020-01-27 a5d6393
fix bug using one band image as input in RasterMath
