### 2020-01-30 c12cdcf
fix bug in mask with rasterMath
