+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fix bug when in RasterMath when input is only one band
+- Fix bug in RasterMath with mask and list
 
 ### Added
 
