### 2020-02-03 cb9d7c6
Changes init_processing
add parallel method 
add write block 
add get coord block
