+++ museotoolbox/processing/__init__.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+ #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # =============================================================================
 # ___  ___                       _____           _______
@@ -23,9 +23,13 @@ import tempfile
 # spatial libraries
 from osgeo import __version__ as osgeo_version
 from osgeo import gdal, ogr
-
+from joblib import Parallel, delayed, cpu_count
 
 from ..internal_tools import ProgressBar, push_feedback
+from time import time, sleep 
+
+from multiprocessing import Pool #  Process pool
+from multiprocessing import sharedctypes
 
 
 def image_mask_from_vector(
@@ -761,7 +765,6 @@ class RasterMath:
             self._init_raster_parameters(compress=compress)
         else:
             params = self.get_raster_parameters()
-            params = self.get_raster_parameters()
             arg_pos = next(
                 (x for x in params if x.startswith('compress')), None)
             if arg_pos:
@@ -1040,6 +1043,65 @@ class RasterMath:
                 tmp = self._manage_2d_mask(tmp)
                 tmp = np.ma.copy(tmp)
             return tmp
+    
+    
+    
+    def get_coord_block(self, block_number=0):
+        """
+        Get position of a block :
+
+            order as [x,y,width,height]
+
+        Parameters
+        -----------
+        block_number, int, optional (default=0).
+            Position of the desired block.
+
+        Returns
+        --------
+        List of positions of the block [x,y,width,height]
+
+        """
+        if block_number > self.n_blocks:
+            raise ValueError(
+                'There are only {} blocks in your image.'.format(
+                    self.n_blocks))
+        else:
+            
+#            for col in range(0, self.n_columns, x_block_size):
+#                width = min(self.n_columns - col, x_block_size)
+#                height = min(self.n_lines - row, y_block_size)
+
+#            if get_block:
+#                X = self._generate_block_array(
+#                    col, row, width, height, self.mask)
+
+
+            row = [l for l in range(0, self.n_lines, self.y_block_size)]
+            col = [c for c in range(0, self.n_columns, self.x_block_size)]
+            
+            row_number = int(block_number / self.n_x_blocks)
+            col_number = int(block_number % self.n_x_blocks)
+            
+            width = min(self.n_columns - col[col_number], self.x_block_size)
+            height = min(self.n_lines - row[row_number], self.y_block_size)
+                        
+            tmp = self._generate_block_array(
+                col[col_number], row[row_number], width, height, self.mask)
+            
+            if self.return_3d is False:
+                tmp = self._manage_2d_mask(tmp)
+                tmp = np.ma.copy(tmp)
+            
+            return [col[col_number], row[row_number], width, height]
+    
+    
+    
+    
+    
+    
+    
+    
 
     def get_random_block(self, random_state=None):
         """
@@ -1210,7 +1272,7 @@ class RasterMath:
                 get_block=True):
 
             if isinstance(X, list):
-                X_ = [np.ma.copy(arr) for arr in X]
+                X_ = [np.ma.copy(arr) for arr in X] #de type numpy mask
                 X = X_[0]  # X_[0] is used to get mask
             else:
                 X_ = np.ma.copy(X)
@@ -1218,13 +1280,13 @@ class RasterMath:
             if self.verbose:
                 self.pb.add_position(self._position)
 
-            for idx, fun in enumerate(self.functions):
+            for idx, fun in enumerate(self.functions):# idx : index et fun les fonctions
                 maxBands = self.outputs[idx].RasterCount
 
                 if not np.all(X.mask == 1):
                     # if all the block is not masked
                     if not self.return_3d:
[…]
