### 2020-02-04 nkarasiak/MuseoToolBox#19 @cmillion
JOSS review, re: software paper
I think that the [paper](https://github.com/nkarasiak/MuseoToolBox/blob/master/paper.md) would benefit from a more specific comparison to the state-of-the-art in the field and how this package improves or differs. It may simply be that the explanation is unclear; the second sentence in the opening paragraph of the paper is very long and difficult to parse. There are also no references provided.

Tagging: openjournals/joss-reviews#1978
