### 2020-02-05 nkarasiak/MuseoToolBox#19 @nkarasiak
Dear Chase,

I'll update the article as asked, but regarding the references, the paper.md file is to generate with JOSS in order to see them : https://github.com/openjournals/joss-papers/blob/joss.01978/joss.01978/10.21105.joss.01978.pdf
[That's what did the awesome whedon](https://github.com/openjournals/joss-reviews/issues/1950#issuecomment-565115635)

Kind regards,
Nicolas.
