### 2020-02-06 03ee95e
Updates
function run_parallel  : 
working but for now  only with unmasked images

delete function parallel_short
