+++ museotoolbox/processing/__init__.py
@@ -1046,7 +1046,7 @@ class RasterMath:
     
     
     
-    def get_coord_block(self, block_number=0):
+    def get_block_coords(self, block_number=0):
         """
         Get position of a block :
 
@@ -1374,93 +1374,85 @@ class RasterMath:
             self.outputs[idx] = None
 
 
-        
-    def run_Parallel (self, n_jobs = 1, verbose_Parallel = 0):
-        """
-        Process writing with outside function with parallel mode using joblib.
-        
-        
-        Parameters
-        ----------
-        n_jobs : int, optional
-            Numbers of workers or process that will work in parallel.
-        verbose_Parallel : int, optional 
-            The verbosity level: if non zero, progress messages are printed.
-            Above 50, the output is sent to stdout. The frequency of the messages 
-            increases with the verbosity level. If it more than 10, all iterations are reported.
 
-        Returns
-        -------
-        None
-        """
-   
-        if n_jobs < 0 :
-            n_jobs = cpu_count()
-        elif n_jobs == 0 :
-            raise ValueError(' {} is not a valid value.'.format(n_jobs))
+    
+    def run_parallel (self, n_jobs = 1, size = '1G'):
+            """
+            Function under construction and used for tests 
+            Apply a function of idx [0] to an image with parallel mode using joblib.
+            Warning : this function does not write the results in an output. 
             
-        test = []
-        self.pb = ProgressBar(self.n_blocks, message=self.message)
-        
-        for i in range(0,self.n_blocks,n_jobs):
-            self.pb.add_position(self._position)
-           
-            if i <=self.n_blocks-n_jobs :
-                idx_blocks=[i for i in range(i,i+n_jobs)]
-            else :
-                idx_blocks = [i for i in range(i,self.n_blocks)]
-            
-            test.append(idx_blocks)
-            for idx,fun in enumerate(self.functions):
-
-                if self.functionsKwargs[idx] is not False:
-                    res=Parallel(n_jobs, verbose = verbose_Parallel)(delayed(fun)(self.get_block(i).data,**self.functionsKwargs[idx])for i in  idx_blocks)
-                else : 
-                    res=Parallel(n_jobs,verbose = verbose_Parallel)(delayed(fun)(self.get_block(i).data)for i in  idx_blocks)
-
-                self._write_few_blocks(idx_blocks , res ,idx)
+            Parameters
+            ----------
+            n_jobs : int, optional, ( default value : 1)
+                Numbers of workers or process that will work in parallel.
+            size : str, optional ( default value : '1G')
+                maximun size of ram the program can use to store temporary the results
                 
-            self._position += 1
-        self.outputs[idx] = None
-        self.pb.add_position(self.n_blocks)
-        
-        return test
+            Returns
+            -------
+            None
+            """
+            #create a list of all the blocks 
+            size_value=size[:-1]
+            if size[-1]=='M':
+                size_value=1048576*int(size_value)
+            elif size[-1]=='G':
+                size_value=1073741824*int(size_value)
+            elif size[-1]=='T':
+                size_value=1099511627776*int(size_value)
+            else : 
+                raise ValueError(' {} is not a valid value. use for example 100G, 10M, 1T '.format(size))
+            tmp=self.get_random_block()
+            length=int(size_value/tmp.size*tmp.itemsize)
+            
+            if length>self.n_blocks:
+                length=self.n_blocks
+           
+            
+            if n_jobs < 0 :
+                n_jobs = cpu_count()
+            elif n_jobs == 0 :
+                raise ValueError(' {} is not a valid value.'.ormat(n_jobs))
+      
+            self.pb = ProgressBar(self.n_blocks, message=self.message)
+
[…]
