### 2020-02-07 5b5995a
Adding parallel to RasterMath
