### 2020-02-09 a420b8c
modify RasterMath, enhance _process_block
