+++ museotoolbox/processing/__init__.py
@@ -1013,13 +1013,14 @@ class RasterMath:
 
         Parameters
         -----------
-        block_number, int, optional (default=0).
+        block_number : int, optional (default=0).
             Position of the desired block.
+        return_with_mask : bool, optinal (default=False)
+            If True and if return_3d is True, will return a numpy masked array.
 
         Returns
         --------
-        Block : np.ndarray
-
+        Block : np.ndarray or np.ma.MaskedArray
         """
         
         if block_number > self.n_blocks:
@@ -1103,8 +1104,8 @@ class RasterMath:
         np.random.seed(random_state)
         rdm = np.random.permutation(np.arange(self.n_blocks))
         idx = 0
-
         size = 0
+        
         while size == 0:
             size = True
             tmp = self.get_block(block_number=rdm[idx])
@@ -1134,12 +1135,10 @@ class RasterMath:
         """
         if x.ndim == 0:
             x = x.reshape(-1, 1)
-        if self.return_3d:
-            if x.ndim == 2:
-                x = x.reshape(*x.shape, 1)
-        else:
-            if x.ndim == 1:
-                x = x.reshape(-1, 1)
+        
+        if x.ndim == self.return_3d+1:
+            x = x.reshape(*x.shape, 1)
+        
         return x
 
     def read_band_per_band(self):
@@ -1161,40 +1160,12 @@ class RasterMath:
                             band == self.nodata, True, False))
                 yield band
 
-    def read_block_per_block(self, x_block_size=False, y_block_size=False):
+    def read_block_per_block(self):
         """
         Yield each block.
         """
-        for X, col, line, cols, lines in self._iter_block(
-                get_block=True, y_block_size=y_block_size, x_block_size=x_block_size):
-            if isinstance(X, list):
-                mask = X[0].mask
-            else:
-                mask = X.mask
-            if not np.all(mask == 1):
-                yield X
-
-    def _return_unmasked_X(self, X):
-        if isinstance(X.mask, np.bool_):
-            if X.mask == False:
-                X = X.data
-            else:
-                pass
-                # no return
-        else:
-            if X.ndim == 1:
-                X = self.reshape_ndim(X)
-            mask = np.in1d(X.mask[:, 0], True)
-            X = X[~mask, :].data
-        return X
-
-    def _manage_2d_mask(self, X):
-        if len(self.opened_images) > 1:
-            X = [self._return_unmasked_X(x) for x in X]
-        else:
-            X = self._return_unmasked_X(X)
-
-        return X
+        for block in range(self.n_blocks):
+            yield self.get_block(block)
 
     def custom_block_size(self, x_block_size=False, y_block_size=False):
         """
@@ -1404,6 +1375,7 @@ class RasterMath:
                     
                 np_dtype :
                     
+                return_3d : 
                 
                 Returns
                 -------
@@ -1411,81 +1383,75 @@ class RasterMath:
                     The block to write
                     
                 """
-                res = True
+                tmp_arr = True
+                
                 if isinstance(block,list):
-                    test_block=block[0]
+                    mask_block=block[0].mask
                 else:
-                    test_block=block
+                    mask_block=block.mask
                     
-                if np.all(test_block.mask == True):
-                    res = False
-                if np.all(test_block.mask == False):
+                # if everything is masked
+                if np.all(mask_block == True):
+                    mask = True
+                    out_block = nodata
+                # if everything is not masked
+                elif np.all(mask_block == False):
                     mask = False
-                elif np.any(test_block.mask == False):
-                    if test_block.ndim == 1:
-                        out_block_shape  = [ test_block.shape[0] ]
-                    else:
-                        out_block_shape = list(test_block.shape[:-1])
-                    res = test_block.size
-                    if isinstance(block,list):
[…]
