### 2020-02-09 6fe12ae
modify RasterMath, enhance _process_block
