+++ examples/processing/rasterMath_testBlockSize_3d_andNBands.py
@@ -33,13 +33,11 @@ image_mask_from_vector(vector,raster,'/tmp/mask.tif',invert=True)
 
 import time
 
-tn = time.time()
-
 for return_3d in [True,False]:
 
     rM = RasterMath(raster,in_image_mask='/tmp/mask.tif',return_3d=return_3d)
     
-    rM.custom_block_size(10,10) # block of 200x200pixels
+    rM.custom_block_size(128,128) # block of  128x128
     
 #    print(rM.get_random_block().shape)
     
