### 2020-02-10 d4a87ad
Enhance paper as recommended by @cmillion, issue #19 Full implementation of parallel, fix issue #18 #21
