+++ museotoolbox/processing/__init__.py
@@ -612,6 +612,7 @@ class RasterMath:
         self.verbose = verbose
         self.message = message
         self.driver = gdal.GetDriverByName('GTiff')
+        self.itemsize = 0
 
         # Load raster
         self.opened_images = []
@@ -629,7 +630,6 @@ class RasterMath:
         # Get block size and itemsize
         band = self.opened_images[0].GetRasterBand(1)
         self.input_block_sizes = band.GetBlockSize()
-        self.itemsize = band.ReadAsArray(0,0,1,1).itemsize*self.n_bands
     
         # input block size
         if block_size is False:
@@ -677,7 +677,7 @@ class RasterMath:
         opened_raster = gdal.Open(in_image, gdal.GA_ReadOnly)
         if opened_raster is None:
             raise ReferenceError('Impossible to open image ' + in_image)
-
+        
         sameSize = True
 
         if len(self.opened_images) > 0:
@@ -686,6 +686,13 @@ class RasterMath:
                 sameSize = False
                 push_feedback("raster {} doesn't have the same size (X and Y) as the initial raster.\n \
                       Museotoolbox can't add it as an input raster.".format(os.path.basename(in_image)))
+        n_bands = opened_raster.RasterCount
+        
+        band = opened_raster.GetRasterBand(1)
+        mem_size = band.ReadAsArray(0,0,1,1).itemsize*n_bands
+        self.itemsize += mem_size
+        del band
+#        self.itemsize += opened_raster.GetRasterBand(0).ReadAsArray(0,0,1,1).itemsize*n_bands 
 
         if sameSize:
             self.opened_images.append(opened_raster)
@@ -797,7 +804,8 @@ class RasterMath:
         self._outputs[-1]['n_bands'] = out_n_bands
         self._outputs[-1]['kwargs'] = kwargs
         self._outputs[-1]['nodata'] = out_nodata
-
+        self._outputs[-1]['itemsize'] = randomBlock.itemsize*out_n_bands
+        
     def _init_raster_parameters(self, compress=True):
 
         self._raster_options = []
@@ -1110,6 +1118,7 @@ class RasterMath:
                 if np.ma.is_masked(tmp[0]):
                     if np.all(tmp[0].mask == True):
                         size = 0
+                
             else:
                 if np.ma.is_masked(tmp):
                     if np.all(tmp.mask == True):
@@ -1224,7 +1233,7 @@ class RasterMath:
         -------
         None
         """
-
+        self._position = 0
         if self.verbose:
             self.pb = ProgressBar(self.n_blocks, message=self.message)
 
@@ -1354,112 +1363,6 @@ class RasterMath:
             -------
             None
             """
-# =============================================================================
-#           Begin_process_block function
-# =============================================================================        
-            def _process_block(fun,kwargs,block,n_bands,nodata,np_dtype,return_3d):
-                """
-                Private function to compute external function per block.
-                In order to save with the size as the input block, this function need the input mask.
-                
-                Parameters
-                ----------
-                fun : 
-                    
-                kwargs :
-                    
-                block :
-                    
-                nodata :
-                    
-                np_dtype :
-                    
-                return_3d : 
-                
-                Returns
-                -------
-                res : np.ndarray
-                    The block to write
-                    
-                """
-                tmp_arr = True
-                
-                if isinstance(block,list):
-                    mask_block=block[0].mask
-                else:
-                    mask_block=block.mask
-                    
-                # if everything is masked
-                if np.all(mask_block == True):
-                    mask = True
-                    out_block = nodata
-                # if everything is not masked
-                elif np.all(mask_block == False):
-                    mask = False
-                
-                # if part masked, part unmasked
-                elif np.any(mask_block == False):
[…]
