### 2020-02-10 nkarasiak/MuseoToolBox#19 @nkarasiak
Hello Chase,
Here is [the new version of the paper](https://github.com/nkarasiak/MuseoToolBox/files/4180641/mtb_0121b2.pdf).
It is generated from the [develop branch](https://github.com/nkarasiak/MuseoToolBox/tree/develop).
Kind regards,
Nicolas.
