+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased : 0.12.1-beta.2]
+## UNRELEASED [0.12.1-beta.3]
+
+### Added
+
+- Spatial functions available for RasterMath
+
+### Changed
+
+- Enhance mask management for RasterMath
+
+## [0.12.1-beta.2] - 2020-02-10
 
 ### Changed
 
