### 2020-02-10 #1978 @kbarnhart
I wanted to check in to keep this review moving along.

I see that there are open issues in the main repository made by @cmillion, Thank you for making issues there.

This ping is meant as a reminder to the reviewers to make any additional issues that stand in the way of completing the review checklist, and as a reminder to @nkarasiak to address issues.

Thanks again to all for contributing to the JOSS review process.
