### 2020-02-11 nkarasiak/MuseoToolBox#25 @mollenburger
dependencies, gdal/osgeo
I'm installing this into a fresh python 3.5 anaconda evironment on mac osx 10.14.6.

Installation as per readme file returned no errors, however when trying to folow an example I couldn't import museotoolbox.

initially, I got an error like "couldn't find osgeo" so I installed gdal via conda, then tried again, got the following:
ImportError: dlopen(//anaconda/envs/ipykernel_py3/lib/python3.5/site-packages/osgeo/_gdal.so, 2): Library not loaded: @rpath/libjpeg.8.dylib
  Referenced from: /anaconda/envs/ipykernel_py3/lib/libgdal.20.dylib
  Reason: image not found

Eventual solution found here: https://hackernoon.com/install-python-gdal-using-conda-on-mac-8f320ca36d90 -- downgrade jpeg from 9 to 8: conda install -f jpeg=8

probably should include gdal as a dependency and maybe flag the jpeg problem on the readme to save someone some time

cc: https://github.com/openjournals/joss-reviews/issues/1978
