### 2020-02-11 nkarasiak/MuseoToolBox#19 @mollenburger
adding here another couple of comments on the software paper:

- the summary is very technical and not understandable by a non-specialist in machine learning and its application to remote sensing/raster data. A sentence or two at the beginning about how Museo is used would be helpful. For example, classification of an RS image by land cover type is a common thing to do--how does Museo make that easier? (e.g. extracting ROIs, transforming rasters/raster bands, fitting various models, performing spatial cross-validation, etc.)

- the figures aren't referred to anywhere in the text, it would be good to refer to Figure 2 in the section on RasterMath

- There's a lot of focus on RasterMath--how is it different from the tools provided in e.g. rasterio or xarray?
