### 2020-02-11 nkarasiak/MuseoToolBox#26 @mollenburger
documentation - importing data for processing
How should data be imported-- it looks like load historical data just assigns pathnames to the variables, so would I just use raster = "~/path/to/raster.tiff" to use my own file?

What file formats are compatible with museo? shapefiles? NetCDF files? Does museo check whether the vector and raster projections match? If I am preparing an ROI file in, say, QGIS and want to use it for processing in Museo, what do I need to know about how to set up that file?

cc: openjournals/joss-reviews#1978
