+++ museotoolbox/processing/__init__.py
@@ -1,4 +1,4 @@
- #!/usr/bin/env python3
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # =============================================================================
 # ___  ___                       _____           _______
@@ -26,7 +26,6 @@ from osgeo import gdal, ogr
 from joblib import Parallel, delayed
 
 from ..internal_tools import ProgressBar, push_feedback
-
 
 def image_mask_from_vector(
         in_vector, in_image, out_image, invert=False, gdt=gdal.GDT_Byte):
@@ -630,7 +629,7 @@ class RasterMath:
         # Get block size and itemsize
         band = self.opened_images[0].GetRasterBand(1)
         self.input_block_sizes = band.GetBlockSize()
-    
+
         # input block size
         if block_size is False:
             self.x_block_size = self.input_block_sizes[0]
@@ -658,7 +657,7 @@ class RasterMath:
 
         # Initialize the output
         self.lastProgress = 0
-        
+
         self._outputs = []
         self._raster_options = {}
 
@@ -677,7 +676,7 @@ class RasterMath:
         opened_raster = gdal.Open(in_image, gdal.GA_ReadOnly)
         if opened_raster is None:
             raise ReferenceError('Impossible to open image ' + in_image)
-        
+
         sameSize = True
 
         if len(self.opened_images) > 0:
@@ -687,12 +686,12 @@ class RasterMath:
                 push_feedback("raster {} doesn't have the same size (X and Y) as the initial raster.\n \
                       Museotoolbox can't add it as an input raster.".format(os.path.basename(in_image)))
         n_bands = opened_raster.RasterCount
-        
+
         band = opened_raster.GetRasterBand(1)
-        mem_size = band.ReadAsArray(0,0,1,1).itemsize*n_bands
+        mem_size = band.ReadAsArray(0, 0, 1, 1).itemsize * n_bands
         self.itemsize += mem_size
         del band
-#        self.itemsize += opened_raster.GetRasterBand(0).ReadAsArray(0,0,1,1).itemsize*n_bands 
+#        self.itemsize += opened_raster.GetRasterBand(0).ReadAsArray(0,0,1,1).itemsize*n_bands
 
         if sameSize:
             self.opened_images.append(opened_raster)
@@ -705,6 +704,7 @@ class RasterMath:
             out_np_dt=False,
             out_nodata=False,
             compress=True,
+            offset=False,
             **kwargs):
         """
         Add function to rasterMath.
@@ -744,15 +744,13 @@ class RasterMath:
             out_np_dt = convert_dt(dtypeName)
             if self.verbose:
                 push_feedback(
-                'Using datatype from numpy table : {}.'.format(dtypeName))
+                    'Using datatype from numpy table : {}.'.format(dtypeName))
         else:
             dtypeName = np.dtype(out_np_dt).name
             out_np_dt = convert_dt(dtypeName)
-        
-        
+
         # get number of bands
         randomBlock = self.reshape_ndim(randomBlock)
-        
 
         out_n_bands = randomBlock.shape[-1]
         need_s = ''
@@ -777,10 +775,9 @@ class RasterMath:
             self._init_raster_parameters(compress=compress)
 
         self._add_output(out_image, out_n_bands, out_np_dt)
-        
+
         if len(kwargs) == 0:
             kwargs = False
-
 
         if (out_nodata is True) or (self.nodata is not None) or (
                 self.mask is not False):
@@ -797,15 +794,16 @@ class RasterMath:
 
             if self.verbose:
                 push_feedback('No data is set to : ' + str(out_nodata))
-        
+
         self._outputs[-1]['gdal_type'] = out_np_dt
[…]
