+++ paper.md
@@ -27,23 +27,27 @@ bibliography: paper.bib
 
 `Museo ToolBox` is a python library dedicated to the processing of images in remote sensing.
 
-This library was initially created to share scikit-learn compatible [spatial cross-validations](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html). Then, we implemented an ``ai`` module to help users to generate stratified spatial or non-spatial cross-validation, with state-of-the-art learning methods including an automatic grid search and standardized data using algorithms from scikit-learn.
+In this domain, classifying land cover type is a common and sometimes complex task, regardless of your level of expertise. Recurring procedures such as extracting Regions Of Interest (raster values from your polygon), computing spectral indices or validate a model with a cross-validation can be difficult to implement.
 
-Another contribution is  ``RasterMath`` which connects your array-compatible function to your raster. For example, as in our field a recurrent usage is to fit a model and predict or to use some functions like one to compute a spectral index, `Museo ToolBox` automatically transforms the raster to match your needs (for learning a model, the user needs an array with one line per pixel and its features as columns). 
+`Museo ToolBox` aims to simplify the whole process by making the main treatments more accessible (extracting of Region Of Interests, fitting a model by using a cross-validation, computing NDVI or various spectral indices, performing any kind of array function to your raster)... 
 
-Museo ToolBox's goal is to make working with raster data very easier for scientists or students and to promote the use of spatial cross-validation.
+The main objective of this library is to ease the transposition of array-like functions into to an image and to promote good practices in machine learning.
 
-A [full documentation is available online on read the docs](http://museotoolbox.readthedocs.io/).
+To make `Museo ToolBox` easier to get started with, a [full documentation with lot of examples is available online on read the docs](http://museotoolbox.readthedocs.io/).
 
-# Museo ToolBox functionnalities
+# Museo ToolBox in details
 
-`Museo ToolBox` is organized into several modules :
+`Museo ToolBox` is organized into several modules (Figure 1) :
 
 - [processing](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.processing.html) : raster and vector processing.
 - [cross-validation](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html) : stratified cross-validation compatible with scikit-learn
-- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) : machine learning module
+- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) :  artificial intelligence module.
 - [charts](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.charts.html) : plot confusion matrix with F1 score, mean, or producer/user's accuracy.
-- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute stats (like Moran's Index, confusion matrix, commision/omission) or extract truth and predicted label from confusion matrix.
+- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute stats (like Moran's Index, confusion matrix, commision/omission) or extract truth and predicted label from a confusion matrix.
+
+![Museo ToolBox schema.](metadata/schema.png)
+
+
 
 Here are some main usages of `Museo ToolBox` :
 
@@ -55,31 +59,34 @@ Here are some main usages of `Museo ToolBox` :
 
 ## RasterMath
 
-Available in `museotoolbox.processing.RasterMath`, RasterMath class is the keystone of ``Museo ToolBox``.
+Available in `museotoolbox.processing`, `RasterMath` class is the keystone of ``Museo ToolBox``.
 
[…]
