+++ museotoolbox/internal_tools/__init__.py
@@ -13,7 +13,6 @@
 # @git:     www.github.com/nkarasiak/MuseoToolBox
 # =============================================================================
 
-
 def push_feedback(msg, feedback=None):
     # in order to convert in Qgis Processing
     # =============================================================================
@@ -28,6 +27,7 @@ def push_feedback(msg, feedback=None):
 
 
 class ProgressBar:
+    
     def __init__(self, total, message='', length=40):
         """
         total : int
@@ -54,20 +54,21 @@ class ProgressBar:
             If false, will add one.
         """
 
-        if value is False:
+        if value :
+            inPercent = int(value / self.total * 100)
+        else:
+    
             self.start += 1
             value = self.start
             inPercent = int(self.start / self.total * 100)
 
-        else:
-            inPercent = int(value / self.total * 100)
         if inPercent != self.lastPosition:
             self.lastPosition = inPercent
             self.nHash = int(self.length * (value / self.total))
             self.nPoints = int(self.length - int(self.nHash))
 
             self.printBar(inPercent)
-
+    
     def printBar(self, value):
         if value == 100:
             end = "\n"
