### 2020-02-17 #1978 @kbarnhart
:wave: another message to check in and keep this review moving forward. I see that both @cmillion and @mollenburger have made issues in the main repo (thanks 👍 ). Based on looking at them it appears that some of the issues may need to be addressed before the reviewers can complete their reviews.

@nkarasiak can I ping you to address these issues so the reviewers can complete their reviews.
