+++ requirements.txt
@@ -3,3 +3,4 @@ scipy>=1.0
 scikit-learn>=0.16
 matplotlib
 joblib
+psutil
