+++ paper.md
@@ -40,9 +40,9 @@ To make `Museo ToolBox` easier to get started with, a [full documentation with l
 `Museo ToolBox` is organized into several modules (Figure 1) :
 
 - [processing](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.processing.html) : raster and vector processing.
-- [cross-validation](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html) : stratified cross-validation compatible with scikit-learn
-- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) :  artificial intelligence module.
-- [charts](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.charts.html) : plot confusion matrix with F1 score, mean, or producer/user's accuracy.
+- [cross-validation](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html) : stratified cross-validation compatible with scikit-learn.
+- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) :  artificial intelligence module built upon scikit-learn.
+- [charts](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.charts.html) : plot confusion matrix with F1 score or producer/user's accuracy.
 - [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute stats (like Moran's Index, confusion matrix, commision/omission) or extract truth and predicted label from a confusion matrix.
 
 ![Museo ToolBox schema.](metadata/schema.png)
@@ -63,19 +63,17 @@ Available in `museotoolbox.processing`, `RasterMath` class is the keystone of ``
 
 The question I asked myself is : How can we make it as easy as possible to implement array-like functions  to images ? The idea behind ``RasterMath`` is, If the function is intended to operate with an array, it should be easy to use it with your raster using as few lines as possible.
 
-So, what does ``RasterMath`` really do ? The user only works with an array an confirms with a sample that the process is doing well, and let `RasterMath` generalizing to the whole image. The user doesn't have to manage the raster reading and writing process, the no-data management, the compression, the number of bands or the projection. Figure 2 explains how Raster Math read a raster, performs your function, and write it to a raster.
+So, what does ``RasterMath`` really do ? The user only works with an array an confirms with a sample that the process is doing well, and let `RasterMath` generalizing to the whole image. The user doesn't have to manage the raster reading and writing process, the no-data management, the compression, the number of bands or the projection. Figure 2 explains how `RasterMath` read a raster, performs your function, and write it to a new raster.
 
-The objective of ``RasterMath`` is to **let the user only focus on his array-compatible function**, and``RasterMath`` manage the raster part. 
+The objective is to **let the user only focus on his array-compatible function**, and``RasterMath`` manages the raster part. 
 
 [See RasterMath documentation and examples](https://museotoolbox.readthedocs.io/en/latest/modules/processing/museotoolbox.processing.RasterMath.html)
 
 ![RasterMath under the hood](metadata/RasterMath_schema.png)
 
-# 
+## Artificial Intelligence
 
-## ai module
-
[…]
