### 2020-02-22 f2dd2e4
fix bug in RasterMath when output is not full path, and enhance run function
