+++ museotoolbox/processing/__init__.py
@@ -623,7 +623,10 @@ class RasterMath:
 
         self.add_image(in_image)
 
-        self.n_jobs = n_jobs
+        if n_jobs < 0:
+            self.n_jobs = os.cpu_count()
+        else:
+            self.n_jobs = n_jobs
         self.n_bands = self.opened_images[0].RasterCount
         self.n_columns = self.opened_images[0].RasterXSize
         self.n_lines = self.opened_images[0].RasterYSize
@@ -883,8 +886,8 @@ class RasterMath:
                 self._raster_options.extend(['TILED=YES'])
         
     def _add_output(self, out_image, out_n_bands, out_np_dt):
-        if not os.path.exists(os.path.dirname(out_image)):
-            os.makedirs(os.path.dirname(out_image))
+        if not os.path.exists(os.path.dirname(os.path.abspath(out_image))):
+            os.makedirs(os.path.dirname(os.path.abspath(out_image)))
 
         self._managed_raster_parameters()
 
@@ -1350,7 +1353,131 @@ class RasterMath:
                             fun['function'].__name__)))
             fun['gdal_object'] = None
 
-    def run(self, memory_size = -1):
+    
+    
+    # =============================================================================
+    #           Begin_process_block function for RasterMath
+    #               Function is outside of class in order to be used with Parallel
+    # =============================================================================        
+    @staticmethod
+    def _process_block(fun,kwargs,block,n_bands,nodata,np_dtype,return_3d,idx_block):
+        """
+        Private function to compute external function per block.
+        In order to save with the size as the input block, this function need the input mask.
+        
+        Parameters
+        ----------
+        fun : function
+            
+        kwargs : kwargs
+            False, or dict of kwargs.
+        block : np.ndarray or np.ma.MaskedArray
+            The function will compute using the given block
+        nodata : nodata, int or float
+            No Data value is some pixels are masked
+        np_dtype : numpy datatype
+            Numpy datatype for the output array
+        return_3d : boolean
+            2d-array or 3d-array.
+        
+        Returns
+        -------
+        res : np.ndarray
+            The block to write
+            
+        """
+        tmp_arr = True
+        
+        if isinstance(block,list):
+            mask_block=block[0].mask
+        else:
+            mask_block=block.mask
+            
+        # if everything is masked
+        if np.all(mask_block == True):
+            mask = True
+            out_block = nodata
+        # if everything is not masked
+        elif np.all(mask_block == False):
+            mask = False
+        
+        # if part masked, part unmasked
+        elif np.any(mask_block == False):
+            tmp_arr = mask_block.size
+            if return_3d:
+                if mask_block.ndim > 2:
+                    mask = mask_block[...,0]
+            elif mask_block.ndim == 1:
+                mask = mask_block
+            else:
+                mask = mask_block[...,0]
+                
+        # if block is not fully masked
+        if mask is not True: 
+                
+            # if no mask, we send the np.ndarray only
+            if mask is False:
+                if isinstance(block,list):
+                    block = [b.data for b in block]
+                else:
+                    block = block.data
+                if kwargs:
+                    out_block = fun(block,**kwargs)
+                else:
+                    out_block = fun(block)
+            
+            # if part is masked
+            else:   
+                # if 3d we send the np.ma.MaskedArray
+                if return_3d:
+                    if kwargs:
+                        out_block = fun(block,**kwargs)
+                    else:
+                        out_block = fun(block)
+                        
+                    if out_block.ndim == 1:
+                        out_block = np.expand_dims(out_block,2)
+                    out_block[mask,...] = nodata
+                
+                # if 2d, we send only the np.ndarray without masked data
+                else:    
+                    # create empty array with nodata value
+                    if mask_block.ndim > 1:
+                        mbs = mask_block.shape[:-1]
+                    else:
[…]
