### 2020-02-22 9b3ef12
fix bug in RasterMath when output is not full path, and enhance run function
