+++ test/test_processing.py
@@ -4,13 +4,35 @@ from shutil import copyfile
 import numpy as np
 from museotoolbox import processing
 from museotoolbox.datasets import load_historical_data
-from osgeo import gdal
-
+from osgeo import gdal, osr
 import os
 
 raster,vector = load_historical_data()
 rM = processing.RasterMath(raster)
 mask = processing.image_mask_from_vector(vector,raster,'/tmp/mask.tif')
+
+
+def create_false_image(array,path):
+    # from https://pcjericks.github.io/py-gdalogr-cookbook/raster_layers.html
+    driver = gdal.GetDriverByName('GTiff')
+    outRaster = driver.Create(path, array.shape[1], array.shape[0], 1, gdal.GDT_Byte)
+    outRaster.SetGeoTransform((0, 10, 0, 0, 0, 10))
+    outband = outRaster.GetRasterBand(1)
+    outband.WriteArray(array)
+    outRasterSRS = osr.SpatialReference()
+    outRasterSRS.ImportFromEPSG(4326)
+    outRaster.SetProjection(outRasterSRS.ExportToWkt())
+    outband.FlushCache()
+
+# create autocorrelated tif
+x = np.zeros((100,100),dtype=int)
+# max autocorr
+x[:50,:] = 1
+x[50:,:] = 2
+
+x_mask = np.random.randint(0,2,[100,100]) # random mask
+create_false_image(x,'/tmp/100x100size.tif')
+
 
 class TestRaster(unittest.TestCase):
     def test_convert_datatype(self):
@@ -38,11 +60,16 @@ class TestRaster(unittest.TestCase):
         assert(processing.convert_dt(gdal.GDT_Int16,to_otb_dt=True) == 'int16')
         assert(processing.convert_dt(np.dtype('float64').name,to_otb_dt=True) == 'double')
         
+        assert(processing._convert_gdal_to_otb_dt(18) == 'cdouble') # if unknow, put cdouble (highest type)
+        
     def _assert_np_gdt(self,in_conv,out_dt):
         assert(processing.convert_dt(in_conv)==out_dt)
         
+        
     def test_gdt_minmax_values(self):
         assert(gdal.GDT_UInt16 == processing.get_gdt_from_minmax_values(500))
+        assert(gdal.GDT_UInt32 == processing.get_gdt_from_minmax_values(max_value=155500))
+        assert(gdal.GDT_Int32 == processing.get_gdt_from_minmax_values(max_value=0,min_value=-75500))
         assert(gdal.GDT_Int16 == processing.get_gdt_from_minmax_values(max_value=1,min_value=-5))
         assert(gdal.GDT_Float32 == processing.get_gdt_from_minmax_values(max_value=2,min_value=-55.55))
         assert(gdal.GDT_Byte == processing.get_gdt_from_minmax_values(max_value=222))
@@ -98,7 +125,11 @@ class TestRaster(unittest.TestCase):
         assert(rM.y_block_size == np.ceil(1/3*rM.n_lines))
         
         rM.add_image(raster)
-        
+        self.assertRaises(ValueError,rM.add_image,'/tmp/100x100size.tif')
+        return_x = lambda x : x[0].astype(np.int16)
+        rM.add_function(return_x,'/tmp/test_double.tif')
+        rM.run()
+        os.remove('/tmp/test_double.tif')
         assert(np.all(rM.get_random_block(random_state=12))== np.all(rM.get_random_block(random_state=12)))
         
         
@@ -117,6 +148,9 @@ class TestRaster(unittest.TestCase):
                 pass
             rM_band.add_function(np.mean,axis=1,out_image='/tmp/mean.tif')
             rM_band.run()
+            
+            self.assertRaises(MemoryError,rM_band.run,'1K')
+            
             assert(idx+1 == rM_band.n_bands)                        
[…]
