### 2020-02-23 6e4cc60
if compression, force BIGTIFF
