### 2020-02-23 345787d
move FlushCache to speedup writing
