### 2020-02-23 4ae7923
move FlushCache to speedup writing
