### 2020-02-26 efc772c
better batch processing for rastermath
