### 2020-02-27 #1978 @nkarasiak
Hello @kbarnhart. I'm dealing with the issues, and I submitted my library to conda in order to answer to @mollenburger to ease the installation process with gdal but it takes some times...
I didn't forget to answers the issues, I just want to do it the nicest way possible :)
I hope end of next week everything will be adressed.
Kind regards,
Nicolas.
