### 2020-02-27 nkarasiak/MuseoToolBox#25 @nkarasiak
Hello @mollenburger,

In order to ease the installation process with conda and to have a working installation, I've submited Museo ToolBox to conda-forge repository : https://github.com/conda-forge/staged-recipes/pull/10863#issuecomment-586718548.
I'm currently awaiting for the forge to include my library.

Kind regards,
Nicolas.
