### 2020-02-27 nkarasiak/MuseoToolBox#26 @nkarasiak
Hello @mollenburger,

As it is specified in the docstring of the _load_historical_data()_ function (https://museotoolbox.readthedocs.io/en/latest/modules/datasets/museotoolbox.datasets.load_historical_data.html), the default behavior only returns the path to the raster and the vector (_return_X_y=False)_.

This is a normal behavior because it simulates how the user will have to process when using his own raster and vector.

If you want only to have the X (features) and y (labels), just write : _load_historical_data(return_X_y=True)_

Museo ToolBox is built upon gdal so it will read every format that gdal can read. I didn't test it with lot of files format (png/tif/gpkg) but it does the job.

For writing raster files, the code is very optimized for geotiff and so only write in this format (as it is specified when using RasterMath _add_function_ : https://museotoolbox.readthedocs.io/en/latest/modules/processing/RasterMath/museotoolbox.processing.RasterMath.add_function.html.

Do you think the function _extract_ROI_ should have more information in the docstring ?
