### 2020-02-27 nkarasiak/MuseoToolBox#19 @nkarasiak
Hello @mollenburger,

I'm curently working with a native English speaker on the rewriting of the paper. Figures of course are now referred in the text.

For RasterMath, the idea is to be very high-end user, so it could have been built upon rasterio but it is not the case. It differs from rasterio because the aim of RasterMath is only to ease the adaptation of a vector function (i.e. function initially programmed for an array) to a raster. There is absolutely no need to ask for opening a raster in writing mode, to define the number of output bands, the projection, the block size, etc...
