+++ paper.md
@@ -27,11 +27,11 @@ bibliography: paper.bib
 
 `Museo ToolBox` is a python library dedicated to the processing of images in remote sensing.
 
-In this domain, classifying land cover type is a common and sometimes complex task, regardless of your level of expertise. Recurring procedures such as extracting Regions Of Interest (raster values from your polygon), computing spectral indices or validate a model with a cross-validation can be difficult to implement.
+In this domain, classifying land cover type is a common and sometimes complex task, regardless of your level of expertise. Recurring procedures such as extracting Regions Of Interest (raster values from your polygon), computing spectral indices or validating a model with a cross-validation can be difficult to implement.
 
-`Museo ToolBox` aims to simplify the whole process by making the main treatments more accessible (extracting of Region Of Interests, fitting a model by using a cross-validation, computing NDVI or various spectral indices, performing any kind of array function to your raster)... 
+`Museo ToolBox` aims at simplifying the whole process by making the main treatments more accessible (extracting of Region Of Interests, fitting a model by using cross-validation, computing NDVI or various spectral indices, performing any kind of array function to the raster, etc).
 
-The main objective of this library is to ease the transposition of array-like functions into to an image and to promote good practices in machine learning.
+The main objective of this library is to facilitate the transposition of array-like functions into an image and to promote good practices in machine learning.
 
 To make `Museo ToolBox` easier to get started with, a [full documentation with lot of examples is available online on read the docs](http://museotoolbox.readthedocs.io/).
 
@@ -43,29 +43,29 @@ To make `Museo ToolBox` easier to get started with, a [full documentation with l
 - [cross-validation](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html) : stratified cross-validation compatible with scikit-learn.
 - [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) :  artificial intelligence module built upon scikit-learn.
 - [charts](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.charts.html) : plot confusion matrix with F1 score or producer/user's accuracy.
-- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute stats (like Moran's Index, confusion matrix, commision/omission) or extract truth and predicted label from a confusion matrix.
+- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute statistics (such as Moran's Index, confusion matrix, commision/omission) or extracting truth and predicted label from a confusion matrix.
 
 ![Museo ToolBox schema.](metadata/schema.png)
 
 
 
-Here are some main usages of `Museo ToolBox` :
+Here are some of the main usages of `Museo ToolBox` :
 
-1. [Read and write a raster block per block using your own function](https://museotoolbox.readthedocs.io/en/latest/modules/processing/museotoolbox.processing.RasterMath.html).
-2. [Generate a cross-validation, including spatial cross-validation](https://museotoolbox.readthedocs.io/en/latest/auto_examples/index.html#cross-validation).
-3. [Fit models with scikit-learn, extract accuracy from each cross-validation fold, and predict raster](https://museotoolbox.readthedocs.io/en/latest/modules/ai/museotoolbox.ai.SuperLearner.html).
-4. [Plot confusion matrix and add f1 score or producer/user accuracy](https://museotoolbox.readthedocs.io/en/latest/modules/charts/museotoolbox.charts.PlotConfusionMatrix.html#museotoolbox.charts.PlotConfusionMatrix).
[…]
