### 2020-03-09 nkarasiak/MuseoToolBox#19 @nkarasiak
Hello @mollenburger  and @cmillion,
Here is the new paper version corrected and proofread by a native english colleague.
Here you will find the new paper : [mtb_beta.pdf](https://github.com/nkarasiak/MuseoToolBox/files/4307825/mtb_beta.pdf)

Tell me if it's okay for you.
Kind regards,
Nicolas.
