### 2020-03-09 #1978 @nkarasiak
Hi @kbarnhart  :hand: ,
I addressed these issues several days ago :
- https://github.com/nkarasiak/MuseoToolBox/issues/25 , I'm waiting for conda-forge to accept Museo ToolBox, but everything has been done to correct the bug @mollenburger encountered (gdal cannot be installed via pip, conda is be a great alternative to avoid this bug).
- https://github.com/nkarasiak/MuseoToolBox/issues/26 , I think everything the reviewer wanted is specified in the docstring, I specified it in the issue.

Today, I updated/corrected the paper according to suggestions from @mollenburger and @cmillion :
- https://github.com/nkarasiak/MuseoToolBox/issues/19

In any case I'm waiting for their approval (by letting them commenting or closing each issue).
