### 2020-03-10 #1978 @kbarnhart
@nkarasiak thanks for the update. Perhaps you can provide another update once the conda-forge PR is merged. This will let the reviewer's know that it is sensible for them to continue with the review process.
