### 2020-03-10 nkarasiak/MuseoToolBox#25 @nkarasiak
Hi @mollenburger,
You can now install museotoolbox directly in conda, so it should be better for you as conda is checking if library is working for each OS.
To install museotoolbox from conda, see here : https://github.com/conda-forge/museotoolbox-feedstock#installing-museotoolbox
Let me know if it works now.

Kind regards,
Nicolas
