### 2020-03-10 #1978 @nkarasiak
Here it is, my first conda-forge package :fireworks:  : https://github.com/conda-forge/museotoolbox-feedstock.
I updated the corresponding issue : https://github.com/nkarasiak/MuseoToolBox/issues/25
All the issues has been addressed, I'm now waiting for reviewers' feedback.
