+++ museotoolbox/processing/__init__.py
@@ -1,4 +1,4 @@
- #!/usr/bin/env python3
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # =============================================================================
 # ___  ___                       _____           _______
@@ -475,6 +475,7 @@ def extract_ROI(in_image, in_vector, *fields, **kwargs):
 
     return toReturn
 
+
 def rasterize(in_image, in_vector, in_field=False, out_image='MEM',
               gdt=gdal.GDT_Int16, invert=False):
     """
@@ -547,6 +548,7 @@ def rasterize(in_image, in_vector, in_field=False, out_image='MEM',
 
     return dst_ds
 
+
 class RasterMath:
     """
     Read one or multiple rasters per block, and perform one or many functions to one or many geotiff raster outputs.
@@ -593,7 +595,7 @@ class RasterMath:
     Saved /tmp/test.tif using function mean
     """
 
-    def __init__(self, in_image, in_image_mask=False, return_3d=False, block_size=[256, 256], 
+    def __init__(self, in_image, in_image_mask=False, return_3d=False, block_size=[256, 256],
                  n_jobs=1, message='rasterMath...', verbose=True):
 
         self.verbose = verbose
@@ -621,7 +623,7 @@ class RasterMath:
         # Get block size and itemsize
         band = self.opened_images[0].GetRasterBand(1)
         self.input_block_sizes = band.GetBlockSize()
-        
+
         # input block size
         if block_size is False:
             self.x_block_size = self.input_block_sizes[0]
@@ -649,7 +651,7 @@ class RasterMath:
 
         # Initialize the output
         self.lastProgress = 0
-        
+
         self._outputs = []
         self._raster_options = {}
 
@@ -668,7 +670,7 @@ class RasterMath:
         opened_raster = gdal.Open(in_image, gdal.GA_ReadOnly)
         if opened_raster is None:
             raise ReferenceError('Impossible to open image ' + in_image)
-        
+
         sameSize = True
 
         if len(self.opened_images) > 0:
@@ -678,12 +680,12 @@ class RasterMath:
                 raise ValueError("raster {} doesn't have the same size (X and Y) as the initial raster.\n \
                       Museotoolbox can't add it as an input raster.".format(os.path.basename(in_image)))
         n_bands = opened_raster.RasterCount
-        
+
         band = opened_raster.GetRasterBand(1)
-        mem_size = band.ReadAsArray(0,0,1,1).itemsize*n_bands
+        mem_size = band.ReadAsArray(0, 0, 1, 1).itemsize * n_bands
         self.itemsize += mem_size
         del band
-#        self.itemsize += opened_raster.GetRasterBand(0).ReadAsArray(0,0,1,1).itemsize*n_bands 
+#        self.itemsize += opened_raster.GetRasterBand(0).ReadAsArray(0,0,1,1).itemsize*n_bands
 
         if sameSize:
             self.opened_images.append(opened_raster)
@@ -735,15 +737,13 @@ class RasterMath:
             out_np_dt = convert_dt(dtypeName)
             if self.verbose:
                 push_feedback(
-                'Using datatype from numpy table : {}.'.format(dtypeName))
+                    'Using datatype from numpy table : {}.'.format(dtypeName))
         else:
             dtypeName = np.dtype(out_np_dt).name
             out_np_dt = convert_dt(dtypeName)
-        
-        
+
         # get number of bands
         randomBlock = self.reshape_ndim(randomBlock)
-        
 
         out_n_bands = randomBlock.shape[-1]
         need_s = ''
@@ -768,10 +768,9 @@ class RasterMath:
             self._init_raster_parameters(compress=compress)
 
[…]
