### 2020-03-19 nkarasiak/MuseoToolBox#27 @cmillion
warning on installation
When installing with `python3 -m pip install museotoolbox --user`, _pip_ throws the following:

> WARNING: The scripts f2py, f2py3 and f2py3.7 are installed in '/home/[user]/.local/bin' which is not on PATH.
>   Consider adding this directory to PATH or, if you prefer to suppress this warning, use --no-warn-script-location.

This might be a non-issue for functionality. I don't know.

openjournals/joss-reviews#1978
