### 2020-03-19 nkarasiak/MuseoToolBox#27 @nkarasiak
hi @cmillion,
Here is an answer :
https://stackoverflow.com/questions/58051467/fixing-path-for-python-libraries-using-bash
If you want to install with f2py support (which is a scipy dependency), just remove the` --user` and add `sudo` before `python3` command.
