### 2020-03-19 nkarasiak/MuseoToolBox#28 @cmillion
"osgeo" dependency missing
Starting from a fresh install of _miniconda_ with this package installed.

`from museotoolbox.ai import SuperLearner`

fails with

> ModuleNotFoundError: No module named 'osgeo'

Which isn't in the requirements.txt.

openjournals/joss-reviews#1978
