### 2020-03-19 nkarasiak/MuseoToolBox#28 @cmillion
"osgeo" is in GDAL, so you this is fixed by installing _gdal_
