### 2020-03-19 #1978 @cmillion
As soon as GDAL is added as a dependency, I'll check off "installation" and that will complete my review.
