### 2020-03-19 nkarasiak/MuseoToolBox#28 @nkarasiak
> Starting from a fresh install of _miniconda_ with this package installed.
>
> `from museotoolbox.ai import SuperLearner`
>
> fails with
>
> > ModuleNotFoundError: No module named 'osgeo'
>
> Which isn't in the requirements.txt.
>
> [openjournals/joss-reviews#1978](https://github.com/openjournals/joss-reviews/issues/1978)

How did you install Museo ToolBox?
As it is not possible to install gdal via pip,  I created a conda feedstock to ease the installation process.
Here it is : https://github.com/conda-forge/museotoolbox-feedstock

I'll update the readme file to detail the gdal part.
Thanks for the feedback.
