### 2020-03-19 nkarasiak/MuseoToolBox#28 @cmillion
In installed with `python3 -m pip install museotoolbox --user` per documentation.
