### 2020-03-19 nkarasiak/MuseoToolBox#28 @nkarasiak
Ok thanks, can you install via conda and tell me if it's okay for you ?
`conda install -c conda-forge museotoolbox`
just uninstall gdal and museotoolbox before.
Sorry for the inconvenience.
