### 2020-03-20 #1978 @kbarnhart
Thanks for the update @cmillion.

@nkarasiak based on @cmillion 's comment I took a look through the repo today. I would recommend you update the readme, documentation, requirements, environment files based on some of the recent changes. For example, I couldn't see anything on the readme or documentation about installing with conda or what the difference between the three requirements files (requirements.txt, requirements-dev.txt, and environment.yml).

@mollenburger please let me know if you have an estimated timeline for completing your review.

Thanks to all for participating in the JOSS review process.
