### 2020-03-21 #1978 @nkarasiak
Hi @kbarnhart, @cmillion  and @mollenburger ,
I updated the [develop branch](https://github.com/nkarasiak/MuseoToolBox/tree/develop) according to all your suggestions.
- The [README.md](https://github.com/nkarasiak/MuseoToolBox/blob/develop/README.md) now details the differences between `conda` and `pip`, and tell the user to install `Museo ToolBox` with `conda` to avoid the `gdal` installation on its own. As specified in the README, `gdal` can not be a `pip` dependency.
- The [CONTRIBUTING.rst](https://github.com/nkarasiak/MuseoToolBox/blob/develop/CONTRIBUTING.rst) now specifies how to install the conda environnement and how to make the local documentation and the python tests.
- The requirements/environment files are now detailed by a comment on top of each file.

Kind regards,
Nicolas.
