+++ .pyup.yml
@@ -1,22 +1,0 @@
-# configure updates globally
-# default: all
-# allowed: all, insecure, False
-update: all
-
-# set the default branch
-# default: empty, the default branch on GitHub
-branch: develop
-
-# search for requirement files
-# default: True
-# allowed: True, False
-search: True
-
-# Specify requirement files by hand, default is empty
-# default: empty
-# allowed: list
-
-requirements:
-  - requirements.txt:
-      pin: False
-  - requirements_dev.txt
