### 2020-03-25 ce016c3
modified get_random_block to get only unmasked array in 2d
