+++ museotoolbox/processing/__init__.py
@@ -227,6 +227,7 @@ def extract_ROI(in_image, in_vector, *fields, **kwargs):
     """
     Extract raster values from Regions Of Interest in a vector file.
     Initially written by Mathieu Fauvel, improved by Nicolas Karasiak.
+
     Parameters
     -----------
     in_image : str.
@@ -238,13 +239,14 @@ def extract_ROI(in_image, in_vector, *fields, **kwargs):
         Each field to extract label/value from.
     **kwargs : list of kwargs.
         - get_pixel_position : bool, optional (default=False).
-        If `get_pixel_position=True`, will return pixel position in the image for each point.
+            If `get_pixel_position=True`, will return pixel position in the image for each point.
         - only_pixel_position : bool, optional (default=False).
-        If `only_pixel_position=True`, with only return pixel position for each point.
+            If `only_pixel_position=True`, with only return pixel position for each point.
         - prefer_memory : bool, optional (default=False).
-        If `prefer_memory=False`, will write temporary raster on disk to extract ROI values.
+            If `prefer_memory=False`, will write temporary raster on disk to extract ROI values.
         - verbose : bool or int, optional (default=True).
-        The higher is the int verbose, the more it will returns informations.
+            The higher is the int verbose, the more it will returns informations.
+
     Returns
     --------
     X : np.ndarray, size of (n_samples,n_features).
@@ -253,9 +255,11 @@ def extract_ROI(in_image, in_vector, *fields, **kwargs):
         Each line of the matrix is a pixel.
     y : np.ndarray, size of (n_samples,).
         The label of each pixel.
+
     See also
     ---------
     museotoolbox.processing.read_vector_values : read field values from vector file.
+
     Examples
     ---------
     >>> from museotoolbox.datasets import load_historical_data
@@ -480,6 +484,7 @@ def rasterize(in_image, in_vector, in_field=False, out_image='MEM',
               gdt=gdal.GDT_Int16, invert=False):
     """
     Rasterize vector to the size of data (raster)
+
     Parameters
     -----------
     in_image : str.
@@ -496,6 +501,7 @@ def rasterize(in_image, in_vector, in_field=False, out_image='MEM',
         gdal datatype.
     invert : bool, optional (default=False).
         if invert is True, polygons will have 0 values in the out_image.
+
     Returns
     --------
      dst_ds : gdal object
@@ -1081,6 +1087,37 @@ class RasterMath:
 
             return [col[col_number], row[row_number], width, height]
 
+    def _manage_block_mask(self,block):
+            
+        if isinstance(block, list):
+            mask_block = block[0].mask
+        else:
+            mask_block = block.mask
+
+        # if everything is masked
+        if np.all(mask_block == True):
+            size = 0
+        # if everything is not masked
+        elif np.all(mask_block == False):
+            size = 1
+
+        # if part masked, part unmasked
+        elif np.any(mask_block == False):
+            size = 1
+
+            if self.return_3d:
+                if mask_block.ndim > 2:
+                    mask = mask_block[..., 0]
+
+            elif mask_block.ndim == 1:
+                mask = mask_block
+            else:
+                mask = mask_block[..., 0]
+
+            if isinstance(block, list) and self.return_3d is False:
+                block = [b[~mask, ...].data for b in block]
+        return block
+        
     def get_random_block(self, random_state=None):
         """
         Get a random block from the raster.
@@ -1099,17 +1136,37 @@ class RasterMath:
         size = 0
 
         while size == 0:
-            size = True
-            tmp = self.get_block(block_number=rdm[idx])
-            if len(self.opened_images) > 1:
-                if np.ma.is_masked(tmp[0]):
-                    if np.all(tmp[0].mask == True):
[…]
