+++ Makefile
@@ -47,7 +47,7 @@ doc :
 	mv README.rst docs/source/
 	cd docs/ && make html
 
-ddoc :
+doc_full :
 	m2r README.md
 	mv README.rst docs/source/
 	rm -rf docs/source/auto_examples/
@@ -62,5 +62,5 @@ aautopep8 :
 	autopep8 -air museotoolbox --jobs -1
 
 pytest :
-	pytest -v  --cov=museotoolbox
+	pytest-3 -v  --cov=museotoolbox
 
