+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED [0.12.1b3]
+## UNRELEASED [0.12.1-rc.1]
 
 ### Added
 - RasterMath use available memory to speed up process and manage now several cores (n_jobs)
