+++ test/test_cross_validation.py
@@ -23,7 +23,18 @@ smallest_class = np.min(np.unique(y,return_counts=True)[1])
 
 
 class TestCV(unittest.TestCase):
-    
+    def test_train_split(self):
+        np.random.seed(42)
+        y = np.random.randint(1,3,10)
+        X = np.random.randint(1,255,[10,3],dtype=np.uint8)
+
+        cv = cross_validation.LeaveOneOut(random_state=42)
+
+        X_train, X_test, y_train, y_test = cross_validation.train_test_split(cv,X,y)
+        assert ( X_train.shape[0]+X_test.shape[0] == X.shape[0] )
+        assert ( y_train.shape[0]+y_test.shape[0] == y.shape[0] )
+        assert ( np.all( np.equal(y_test,np.array([1,2]) ) ) )
+
     def test_loo(self):
         for split in [False,1,2,5]:
             
