+++ examples/processing/rasterMathCustomBlockAndMask.py
@@ -34,7 +34,7 @@ raster,vector = datasets.load_historical_data()
 
 image_mask_from_vector(vector,raster,'/tmp/mask.tif',invert=False)
 
-rM = RasterMath(raster,in_image_mask='q',return_3d=True)
+rM = RasterMath(raster,in_image_mask='/tmp/mask.tif',return_3d=True)
 #rM.addInputRaster('/tmp/mask.tif')
 print(rM.get_random_block().shape)
 
