+++ .environment.yml
@@ -2,6 +2,12 @@
 
 name: museotoolbox-dev
 dependencies:
+	# ReadTheDoc dependencies
+	- mock
+	- pillow
+	- sphinx
+	- sphinx_rtd_theme
+	# other dependencies
         - python=3.7
         - pytest
         - gdal
