### 2020-04-18 bc5f8bd
update .environment.yml for rtd
