+++ .environment.yml
@@ -1,11 +1,15 @@
 # Conda environment for developpers
 
 name: museotoolbox-dev
+channels:
+- conda-forge
+- defaults
+
 dependencies:
 	# ReadTheDoc dependencies
 	- mock
 	- pillow
-	- sphinx
+	- sphinx==2.2.1
 	- sphinx_rtd_theme
 	# other dependencies
         - python=3.7
