### 2020-04-18 65970df
update .environment.yml for rtd
