### 2020-04-18 63f07c3
update .environment.yml for rtd
