+++ .environment.yml
@@ -3,10 +3,6 @@
 name: museotoolbox-dev
 
 dependencies:
-	- mock
-	- pillow
-	- sphinx==2.2.1
-	- sphinx_rtd_theme
         - python=3.7
         - pytest
         - gdal
