### 2020-04-19 #1978 @kbarnhart
@nkarasiak here are a few editorial comments on the paper. These comments mostly pertain to grammar, acronyms, references, and making sure the paper is suitable for a general audience. Once these are addressed, let me know and we will move forward with the manuscript process.

- [ ] I think the Summary section would benefit from a statement describing what a raster is relative to, say, and image or an array. This statement would make the remainder of the paper more accessible for a general audience.
- [ ] Define the acronym NDVI when first used.
- [ ] Provide a reference for Moran's I
- [ ] Provide a reference for scikit-learn
- [ ] There are extra spaces between text and question marks and text and colons in the main text.
- [ ] "How can we make it as easy as possible to implement array-like functions to images?" should be "How can we make it as easy as possible to implement array-like functions on images ?" (to -> on) In addition, I note that this section will make more sense to a general audience if a statement of what an "image" is and a "raster" is can be found in the summary.
- [ ] "The user only works with an array an confirms with a sample that the process is doing well," should be "The user only works with an array and confirms with a sample that the process is doing well," (an -> and)
- [ ] "the number of bands or the projection" should be "the number of bands, or the projection" (addition of a comma after bands)
- [ ] "The artificial intelligence (ai) module is natively built to implement scikit-learn algorithm" should be "The artificial intelligence (ai) module is natively built to implement scikit-learn algorithms" (algorithm -> algorithms)
- [ ] The three references in the section on Cross-validation should be in-text instead of parenthetical. "As stated by (Olofsson et al., 2014)" should use an in-text citation.
- [ ] There are four undefined references in the Acknowledgments (perhaps you are calling out GitHub handles, but it is not clear)
- [ ] "For spatial cross-validation,
see (Karasiak et al., 2019) inspired from (Roberts et al., 2017)" should be "For spatial cross-validation, see Karasiak et al. (2019) inspired by Roberts et al. (2017)" (from -> by)
