### 2020-04-20 7f27668
update paper for JOSS
