+++ paper.bib
@@ -1,3 +1,31 @@
+@article{moran_notes_1950,
+	title = {Notes on {Continuous} {Stochastic} {Phenomena}},
+	volume = {37},
+	issn = {0006-3444},
+	url = {http://www.jstor.org/stable/2332142},
+	doi = {10.2307/2332142},
+	number = {1/2},
+	urldate = {2017-12-08},
+	journal = {Biometrika},
+	author = {Moran, P. A. P.},
+	year = {1950},
+	pages = {17--23}
+}
+
+@article{scikitlearn_2011,
+author = {Pedregosa, Fabian and Varoquaux, Ga\"{e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, \'{E}douard},
+title = {Scikit-Learn: Machine Learning in Python},
+year = {2011},
+issue_date = {February 2011},
+publisher = {JMLR.org},
+volume = {12},
+number = {null},
+issn = {1532-4435},
+journal = {J. Mach. Learn. Res.},
+month = nov,
+pages = {2825–2830},
+numpages = {6}
+}
 @article{karasiak_2019,
 	title = {Statistical stability and spatial unstability in prediction of forest tree species using satellite image time series},
 	journal = {Remote Sensing},
