+++ paper.md
@@ -25,11 +25,11 @@ bibliography: paper.bib
 
 # Summary
 
-`Museo ToolBox` is a python library dedicated to the processing of images in remote sensing.
+`Museo ToolBox` is a python library dedicated to the processing of images in remote sensing. Images, also known as rasters, are georeferenced arrays.
 
-In this domain, classifying land cover type is a common and sometimes complex task, regardless of your level of expertise. Recurring procedures such as extracting Regions Of Interest (raster values from your polygon), computing spectral indices or validating a model with a cross-validation can be difficult to implement.
+In remote sensing, classifying land cover type is a common and sometimes complex task, regardless of your level of expertise. Recurring procedures such as extracting Regions Of Interest (raster values from your polygon), computing spectral indices or validating a model with a cross-validation can be difficult to implement.
 
-`Museo ToolBox` aims at simplifying the whole process by making the main treatments more accessible (extracting of Region Of Interests, fitting a model by using cross-validation, computing NDVI or various spectral indices, performing any kind of array function to the raster, etc).
+`Museo ToolBox` aims at simplifying the whole process by making the main treatments more accessible (extracting of Region Of Interests, fitting a model by using cross-validation, computing Normalized Difference Vegetation Index (NDVI) or various spectral indices, performing any kind of array function to the raster, etc).
 
 The main objective of this library is to facilitate the transposition of array-like functions into an image and to promote good practices in machine learning.
 
@@ -41,9 +41,9 @@ To make `Museo ToolBox` easier to get started with, a [full documentation with l
 
 - [processing](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.processing.html) : raster and vector processing.
 - [cross-validation](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html) : stratified cross-validation compatible with scikit-learn.
-- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) :  artificial intelligence module built upon scikit-learn.
+- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) :  artificial intelligence module built upon scikit-learn [@scikitlearn_2011].
 - [charts](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.charts.html) : plot confusion matrix with F1 score or producer/user's accuracy.
-- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute statistics (such as Moran's Index, confusion matrix, commision/omission) or extracting truth and predicted label from a confusion matrix.
+- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute statistics (such as Moran's Index [@moran_notes_1950], confusion matrix, commision/omission) or extracting truth and predicted label from a confusion matrix.
 
 ![Museo ToolBox schema.](metadata/schema.png)
 
@@ -61,11 +61,11 @@ Here are some of the main usages of `Museo ToolBox` :
 
 Available in `museotoolbox.processing`, `RasterMath` class is the keystone of ``Museo ToolBox``.
 
-The question I asked myself is : How can we make it as easy as possible to implement array-like functions  to images ? The idea behind ``RasterMath`` is that if the function is intended to operate with an array, it should be easy to use it with your raster using as few lines as possible.
+The question I asked myself is : How can we make it as easy as possible to implement array-like functions  to images? The idea behind ``RasterMath`` is that if the function is intended to operate with an array, it should be easy to use it with your raster using as few lines as possible.
 
[…]
