### 2020-04-20 2d61c21
update paper for JOSS
