### 2020-04-20 #1978 @nkarasiak
Hi @kbarnhart,
Thank you for the feedback, I updated according to your suggestions. The updated paper is only on the develop branch, let me know if I have to move it to the master branch.

Kind regards,
Nicolas.
