### 2020-04-20 #1978 @kbarnhart
@nkarasiak it looks to me like there are a couple of changes still remaining.

Once you have made them, please merge these changes into the main (master) branch.
