### 2020-04-20 cab1a17
update paper for JOSS
