### 2020-04-20 #1978 @nkarasiak
Hi @kbarnhart, sorry for the mistakes.
- I checked all the double spaces and extra spaces before question marks.
- I updated the two sentences to enhance you identified.
