### 2020-04-20 #1978 @kbarnhart
@nkarasiak not a problem. There are, however, still some  extra spaces before colons in the text.
