### 2020-04-20 cb473f2
update paper for JOSS
