+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Museo ToolBox : a python library for remote sensing including a new way to handle rasters.'
+title: 'Museo ToolBox: a python library for remote sensing including a new way to handle rasters.'
 
 tags:
   - Python
@@ -37,19 +37,19 @@ To make `Museo ToolBox` easier to get started with, a [full documentation with l
 
 # Museo ToolBox in details
 
-`Museo ToolBox` is organized into several modules (Figure 1) :
+`Museo ToolBox` is organized into several modules (Figure 1):
 
-- [processing](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.processing.html) : raster and vector processing.
-- [cross-validation](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html) : stratified cross-validation compatible with scikit-learn.
-- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html) : artificial intelligence module built upon scikit-learn [@scikitlearn_2011].
-- [charts](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.charts.html) : plot confusion matrix with F1 score or producer/user's accuracy.
-- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html) : compute statistics (such as Moran's Index [@moran_notes_1950], confusion matrix, commision/omission) or extracting truth and predicted label from a confusion matrix.
+- [processing](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.processing.html): raster and vector processing.
+- [cross-validation](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.cross_validation.html): stratified cross-validation compatible with scikit-learn.
+- [ai](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.ai.html): artificial intelligence module built upon scikit-learn [@scikitlearn_2011].
+- [charts](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.charts.html): plot confusion matrix with F1 score or producer/user's accuracy.
+- [stats](https://museotoolbox.readthedocs.io/en/latest/modules/museotoolbox.stats.html): compute statistics (such as Moran's Index [@moran_notes_1950], confusion matrix, commision/omission) or extracting truth and predicted label from a confusion matrix.
 
 ![Museo ToolBox schema.](metadata/schema.png)
 
 
 
-Here are some of the main usages of `Museo ToolBox` :
+Here are some of the main usages of `Museo ToolBox`:
 
 1. [Reading and writing a raster block per block using your own function](https://museotoolbox.readthedocs.io/en/latest/modules/processing/museotoolbox.processing.RasterMath.html).
 2. [Generating cross-validation, including spatial cross-validation](https://museotoolbox.readthedocs.io/en/latest/auto_examples/index.html#cross-validation).
@@ -61,7 +61,7 @@ Here are some of the main usages of `Museo ToolBox` :
 
 Available in `museotoolbox.processing`, `RasterMath` class is the keystone of ``Museo ToolBox``.
 
-The question I asked myself is : How can we make it as easy as possible to implement array-like functions on images? The idea behind ``RasterMath`` is that if the function is intended to operate with an array, it should be easy to use it with your raster using as few lines as possible.
+The question I asked myself is: How can we make it as easy as possible to implement array-like functions on images? The idea behind ``RasterMath`` is that if the function is intended to operate with an array, it should be easy to use it with your raster using as few lines as possible.
 
[…]
