### 2020-04-20 #1978 @nkarasiak
@kbarnhart, it seems like I am a big fan of spaces before colons... including in the title... But now I'm cured :)
