+++ paper.bib
@@ -19,9 +19,8 @@ year = {2011},
 issue_date = {February 2011},
 publisher = {JMLR.org},
 volume = {12},
-number = {null},
 issn = {1532-4435},
-journal = {J. Mach. Learn. Res.},
+journal = {Journal of Machine Learning Research},
 month = nov,
 pages = {2825–2830},
 numpages = {6}
