+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Museo ToolBox: a python library for remote sensing including a new way to handle rasters.'
+title: 'Museo ToolBox: A Python library for remote sensing including a new way to handle rasters.'
 
 tags:
   - Python
@@ -25,11 +25,11 @@ bibliography: paper.bib
 
 # Summary
 
-`Museo ToolBox` is a python library dedicated to the processing of georeferenced arrays, also known as rasters or images in remote sensing.
+`Museo ToolBox` is a Python library dedicated to the processing of georeferenced arrays, also known as rasters or images in remote sensing.
 
-In this domain, classifying land cover type is a common and sometimes complex task, regardless of your level of expertise. Recurring procedures such as extracting Regions Of Interest (raster values from your polygon), computing spectral indices or validating a model with a cross-validation can be difficult to implement.
+In this domain, classifying land cover type is a common and sometimes complex task, regardless of your level of expertise. Recurring procedures such as extracting Regions Of Interest (ROIs, raster values from a polygon), computing spectral indices or validating a model with a cross-validation can be difficult to implement.
 
-`Museo ToolBox` aims at simplifying the whole process by making the main treatments more accessible (extracting of Region Of Interests, fitting a model by using cross-validation, computing Normalized Difference Vegetation Index (NDVI) or various spectral indices, performing any kind of array function to the raster, etc).
+`Museo ToolBox` aims at simplifying the whole process by making the main treatments more accessible (extracting of ROIs, fitting a model with cross-validation, computing Normalized Difference Vegetation Index (NDVI) or various spectral indices, performing any kind of array function to the raster, etc).
 
 The main objective of this library is to facilitate the transposition of array-like functions into an image and to promote good practices in machine learning.
 
@@ -47,9 +47,7 @@ To make `Museo ToolBox` easier to get started with, a [full documentation with l
 
 ![Museo ToolBox schema.](metadata/schema.png)
 
-
-
-Here are some of the main usages of `Museo ToolBox`:
+The main usages of `Museo ToolBox` are:
 
 1. [Reading and writing a raster block per block using your own function](https://museotoolbox.readthedocs.io/en/latest/modules/processing/museotoolbox.processing.RasterMath.html).
 2. [Generating cross-validation, including spatial cross-validation](https://museotoolbox.readthedocs.io/en/latest/auto_examples/index.html#cross-validation).
@@ -59,11 +57,11 @@ Here are some of the main usages of `Museo ToolBox`:
 
 ## RasterMath
 
-Available in `museotoolbox.processing`, `RasterMath` class is the keystone of ``Museo ToolBox``.
+Available in `museotoolbox.processing`, the `RasterMath` class is the keystone of ``Museo ToolBox``.
 
 The question I asked myself is: How can we make it as easy as possible to implement array-like functions on images? The idea behind ``RasterMath`` is that if the function is intended to operate with an array, it should be easy to use it with your raster using as few lines as possible.
 
-So, what does ``RasterMath`` really do? The user only works with an array and confirms with a sample that the process is doing well, and lets `RasterMath` generalize it to the whole image. The user doesn't need to manage the raster reading and writing process, the no-data management, the compression, the number of bands, or the projection. Figure 2 explains how `RasterMath` reads a raster, performs the function, and writes it to a new raster.
[…]
