### 2020-04-21 #1978 @kbarnhart
@nkarasiak once you have considered nkarasiak/MuseoToolBox/pull/29 and updated the repositories `master` branch could you:

- [ ] Make a tagged release of your software, and indicate the version tag of the archived version here.
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository).
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
- [ ] Please list the DOI of the archived version here
