### 2020-04-21 c19508e
v0.13.0, suitable for JOSS
