+++ test/test_processing.py
@@ -139,7 +139,9 @@ class TestRaster(unittest.TestCase):
             processing.image_mask_from_vector(vector,raster,mask)
             mask_src = gdal.Open(mask)
             raster_src = gdal.Open(raster)
-            assert(mask_src.GetProjection() == raster_src.GetProjection())
+            mask_proj =osr.SpatialReference(wkt=mask_src.GetProjection())
+            raster_proj = osr.SpatialReference(wkt=raster_src.GetProjection())
+            assert(raster_proj.GetAttrValue('projcs') == mask_proj.GetAttrValue('projcs'))
             assert(mask_src.RasterCount == 1)
             assert(mask_src.RasterXSize == raster_src.RasterXSize)
             assert(mask_src.RasterYSize == raster_src.RasterYSize)
