### 2020-04-21 #1978 @nkarasiak
Hi @kbarnhart
I updated and published the 0.13.0 version in the `master` branch :
- The git tagged version is [0.13.0](https://github.com/nkarasiak/MuseoToolBox/releases/tag/v0.13.0).
- The Zenodo archive version is also 0.13.0 with the [DOI 10.5281/zenodo.3759215](https://zenodo.org/record/3759215). Author list, title and ORCID have been added  in the archive deposit.
