### 2020-04-21 #1978 @kbarnhart
@nkarasiak I'm going to move this article to the final stage by notifying the JOSS editors in chief that it is ready to be published. They may or may note have have additional requests of you.
